refactor app
